refactor context-builder
