+++ packages/model-runtime/src/core/contextBuilders/openai.ts
@@ -1,9 +1,9 @@
 import OpenAI, { toFile } from 'openai';
 
-import { disableStreamModels, systemToUserModels } from '../const/models';
-import { ChatStreamPayload, OpenAIChatMessage } from '../types';
-import { imageUrlToBase64 } from './imageToBase64';
-import { parseDataUri } from './uriParser';
+import { disableStreamModels, systemToUserModels } from '../../const/models';
+import { ChatStreamPayload, OpenAIChatMessage } from '../../types';
+import { imageUrlToBase64 } from '../../utils/imageToBase64';
+import { parseDataUri } from '../../utils/uriParser';
 
 export const convertMessageContent = async (
   content: OpenAI.ChatCompletionContentPart,
