+++ packages/model-runtime/src/core/openaiCompatibleFactory/createImage.ts
@@ -6,8 +6,8 @@ import OpenAI from 'openai';
 import { CreateImagePayload, CreateImageResponse } from '../../types/image';
 import { getModelPricing } from '../../utils/getModelPricing';
 import { imageUrlToBase64 } from '../../utils/imageToBase64';
-import { convertImageUrlToFile } from '../../utils/openaiHelpers';
 import { parseDataUri } from '../../utils/uriParser';
+import { convertImageUrlToFile } from '../contextBuilders/openai';
 import { convertOpenAIImageUsage } from '../usageConverters/openai';
 
 const log = createDebug('lobe-image:openai-compatible');
