+++ packages/model-runtime/src/core/streams/cloudflare.test.ts
@@ -1,12 +1,8 @@
 // @vitest-environment node
 import { Mock, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import * as desensitizeTool from '../utils/desensitizeUrl';
-import {
-  CloudflareStreamTransformer,
-  desensitizeCloudflareUrl,
-  fillUrl,
-} from './cloudflareHelpers';
+import * as desensitizeTool from '../../utils/desensitizeUrl';
+import { CloudflareStreamTransformer, desensitizeCloudflareUrl, fillUrl } from './cloudflare';
 
 afterEach(() => {
   vi.restoreAllMocks();
