+++ packages/model-runtime/src/index.ts
@@ -1,4 +1,5 @@
 export * from './core/BaseAI';
+export { pruneReasoningPayload } from './core/contextBuilders/openai';
 export { ModelRuntime } from './core/ModelRuntime';
 export { createOpenAICompatibleRuntime } from './core/openaiCompatibleFactory';
 export * from './core/RouterRuntime';
@@ -36,5 +37,4 @@ export * from './types/error';
 export { AgentRuntimeError } from './utils/createError';
 export { getModelPropertyWithFallback } from './utils/getFallbackModelProperty';
 export { getModelPricing } from './utils/getModelPricing';
-export { pruneReasoningPayload } from './utils/openaiHelpers';
 export { parseDataUri } from './utils/uriParser';
